Split Challenges into two components
